Add Firebase Core dependency and configuration files
This commit adds the Firebase Core dependency to the project and configures it for both Android and iOS platforms. It also includes the necessary configuration files for Firebase, including `google-services.json` and `GoogleService-Info.plist`. Additionally, a `firebase_options.dart` file is added to provide default Firebase options for the app. Finally, the necessary changes are made to the `main.dart` file to initialize Firebase and the `GeneratedPluginRegistrant` files for Android, iOS, and Windows are updated to include the Firebase Core plugin.