Add Google Sign-In functionality
This commit adds the ability for users to sign in with their Google accounts using the Google Sign-In package. The `FireAuthServices` class now includes a `signInWithGoogle` method that returns a `User` object upon successful sign-in. The `AuthRepoImpl` class implements a `signInWithGoogle` method that returns a `UserEntity` object. The `LoginCubit` class now includes a `loginWithGoogle` method that triggers the sign-in process and updates the state accordingly. A new `LoginByGoogle` widget has been added to allow users to sign in with their Google accounts. Additionally, the necessary dependencies have been added to the `pubspec.yaml` file.