Add Info Circle Icon and Account Helping Section
This commit introduces a new SVG icon, `info_circle_icon.svg`, to the project, along with its corresponding asset declaration in `assets.dart`. Additionally, it includes the implementation of a new section in the account feature, enhancing the user interface with an "Account Helping Section" that provides information about the application or service. This section is composed of a title, a list tile for "من نحن" (Who We Are), and a logout button. The user interface updates also include a new background image for the user profile in the account section. These changes aim to improve user engagement and provide clearer navigation and information within the account settings.